Provide literal data with `I()` for readr 2.0.0 support
In readr 2.0.0 you need to explicitly pass literal data with `I()`.
Previous versions of readr interpreted character vectors with length
greater than 1 as literal data, but readr 2.0.0 now treats these as
multiple filenames to read.

The current package examples fail with readr 2.0.0 due to this, so these
changes will have to be submitted to CRAN to avoid breaking.